Move GetItemFilePath to pkg/archive for shared use
