+++ pkg/restore/restore.go
@@ -22,7 +22,6 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
-	"path/filepath"
 	"sort"
 	"strings"
 	"sync"
@@ -414,7 +413,7 @@ func (ctx *restoreContext) execute() (Result, Result) {
 			// create a blank one.
 			if namespace != "" && !existingNamespaces.Has(targetNamespace) {
 				logger := ctx.log.WithField("namespace", namespace)
-				ns := getNamespace(logger, getItemFilePath(ctx.restoreDir, "namespaces", "", namespace), targetNamespace)
+				ns := getNamespace(logger, archive.GetItemFilePath(ctx.restoreDir, "namespaces", "", namespace), targetNamespace)
 				if _, err := kube.EnsureNamespaceExistsAndIsReady(ns, ctx.namespaceClient, ctx.resourceTerminatingTimeout); err != nil {
 					errs.AddVeleroError(err)
 					continue
@@ -467,15 +466,6 @@ func (ctx *restoreContext) execute() (Result, Result) {
 	ctx.log.Info("Done waiting for all restic restores to complete")
 
 	return warnings, errs
-}
-
-func getItemFilePath(rootDir, groupResource, namespace, name string) string {
-	switch namespace {
-	case "":
-		return filepath.Join(rootDir, velerov1api.ResourcesDir, groupResource, velerov1api.ClusterScopedDir, name+".json")
-	default:
-		return filepath.Join(rootDir, velerov1api.ResourcesDir, groupResource, velerov1api.NamespaceScopedDir, namespace, name+".json")
-	}
 }
 
 // getNamespace returns a namespace API object that we should attempt to
@@ -693,7 +683,7 @@ func (ctx *restoreContext) restoreResource(resource, targetNamespace, originalNa
 	groupResource := schema.ParseGroupResource(resource)
 
 	for _, item := range items {
-		itemPath := getItemFilePath(ctx.restoreDir, resource, originalNamespace, item)
+		itemPath := archive.GetItemFilePath(ctx.restoreDir, resource, originalNamespace, item)
 
 		obj, err := ctx.unmarshal(itemPath)
 		if err != nil {
@@ -784,7 +774,7 @@ func (ctx *restoreContext) restoreItem(obj *unstructured.Unstructured, groupReso
 		// if the namespace scoped resource should be restored, ensure that the namespace into
 		// which the resource is being restored into exists.
 		// This is the *remapped* namespace that we are ensuring exists.
-		nsToEnsure := getNamespace(ctx.log, getItemFilePath(ctx.restoreDir, "namespaces", "", obj.GetNamespace()), namespace)
+		nsToEnsure := getNamespace(ctx.log, archive.GetItemFilePath(ctx.restoreDir, "namespaces", "", obj.GetNamespace()), namespace)
 		if _, err := kube.EnsureNamespaceExistsAndIsReady(nsToEnsure, ctx.namespaceClient, ctx.resourceTerminatingTimeout); err != nil {
 			errs.AddVeleroError(err)
 			return warnings, errs
@@ -968,7 +958,7 @@ func (ctx *restoreContext) restoreItem(obj *unstructured.Unstructured, groupReso
 		obj = unstructuredObj
 
 		for _, additionalItem := range executeOutput.AdditionalItems {
-			itemPath := getItemFilePath(ctx.restoreDir, additionalItem.GroupResource.String(), additionalItem.Namespace, additionalItem.Name)
+			itemPath := archive.GetItemFilePath(ctx.restoreDir, additionalItem.GroupResource.String(), additionalItem.Namespace, additionalItem.Name)
 
 			if _, err := ctx.fileSystem.Stat(itemPath); err != nil {
 				ctx.log.WithError(err).WithFields(logrus.Fields{
